tests/wpt/tests/css/css-backgrounds/parsing/border-width-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing border-width with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-width">
<meta name="assert" content="border-width supports the full grammar '<line-width>{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("border-width", "1em");
test_valid_value("border-width", "2px thin medium thick");

test_valid_value("border-top-width", "0px");
test_valid_value("border-right-width", "thin");
test_valid_value("border-bottom-width", "medium");
test_valid_value("border-left-width", "thick");
</script>
</body>
</html>
